+++ NEWS.md
@@ -4,7 +4,7 @@ This page describes the most important changes in `AstroFITS`. The format is bas
 Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic
 Versioning](https://semver.org).
 
-## Unreleased
+## Version 1.1.0 [2026-01-22]
 
 ### Added